whisper: move the e2e test of PR #15165 to whisper_test.go
whisper_test.go is for general tests while filter_test.go is for
testing filters. The goal of these end to end tests are to be more
generalist and catch a wider range of issues.
